requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
boto3==1.12.39
botocore==1.15.39
colorama==0.4.3
Django==3.0.2
django-admin-honeypot==1.1.0
django-bootstrap-form==3.4
django-bootstrap-modal-forms==1.5.0
django-filter==2.2.0
django-storages==1.9.1
django-widget-tweaks==1.4.5
docutils==0.15.2
gunicorn==20.0.4
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
